# e-commerce_backend/openapi-spec.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/v1/admin/orders/{id}/status": {
            "patch": {
                "operationId": "AdminController_updateOrderStatus",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateOrderStatusDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Admin"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
...
        "/api/v1/orders": {
            "post": {
                "operationId": "OrderController_createOrder",
                "summary": "Create a new order",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully"
                    },
                    "400": {
                        "description": "Invalid order data"
                    },
                    "403": {
                        "description": "Authentication required"
                    }
                },
                "tags": [
                    "Orders"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            },
            "get": {
                "operationId": "OrderController_getAllOrders",
                "parameters": [
                    {
                        "name": "page",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Orders"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/orders/from-cart": {
            "post": {
                "operationId": "OrderController_createOrderFromCart",
                "summary": "Create order from shopping cart",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOrderFromCartDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created from cart"
                    },
                    "400": {
                        "description": "Invalid cart data"
                    }
                },
                "tags": [
                    "Orders"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/orders/stats": {
            "get": {
                "operationId": "OrderController_getUserOrderStats",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Orders"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/orders/seller/financials": {
            "get": {
                "operationId": "OrderController_getSellerFinancials",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Orders"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/orders/seller/orders": {
            "get": {
                "operationId": "OrderController_getSellerOrders",
                "parameters": [
                    {
                        "name": "page",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Orders"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/orders/{id}": {
            "get": {
                "operationId": "OrderController_getOrder",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Orders"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/orders/{id}/status": {
            "patch": {
                "operationId": "OrderController_updateOrderStatus",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateOrderStatusDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Orders"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
        "/api/v1/orders/{id}/cancel": {
            "post": {
                "operationId": "OrderController_cancelOrder",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Orders"
                ],
                "security": [
                    {
                        "JWT-auth": []
                    }
                ]
            }
        },
...
        "/api/v1/image-upload/pic/{imagename}": {
            "get": {
                "operationId": "ImageUploadController_getImage",
                "parameters": [
                    {
                        "name": "imagename",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Image Upload"
                ]
            }
        },
...
    },
    "info": {
        "title": "E-Commerce Platform API",
        "description": "Complete API documentation for the E-Commerce platform with authentication, products, orders, notifications, and seller management.",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [
        {
            "url": "http://localhost:4002",
            "description": "Development Server"
        },
        {
            "url": "http://localhost:4002/api/v1",
            "description": "API v1"
        }
    ],
    "components": {
        "securitySchemes": {
            "JWT-auth": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http",
                "description": "Enter JWT token"
            },
            "cookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "access_token",
                "description": "JWT token stored in HTTP-only cookie"
            }
        },
        "schemas": {
...
            "UpdateOrderStatusDto": {
                "type": "object",
                "properties": {}
            },
...
            "CreateOrderDto": {
                "type": "object",
                "properties": {}
            },
            "CreateOrderFromCartDto": {
                "type": "object",
                "properties": {}
            },
...
        }
    }
}